requirements.txt
discord.py==0.16.4
slixmpp==1.2.3
PyYAML
